+++ CHANGELOG.md
@@ -43,3 +43,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - [#108](https://github.com/babylonlabs-io/babylon-sdk/pull/108) chore: Upgrade wasmd to 0.55
 - [#116](https://github.com/babylonlabs-io/babylon-sdk/pull/116) add docs for
   the x/babylon module
+- [#143](https://github.com/babylonlabs-io/babylon-sdk/pull/143) Re-enable e2e tests